Move suitable codes to elastic-job-lite-tracing-api
